+++ packages/condition_aggregator_digitraffic/lib/condition_aggregator_digitraffic.dart
@@ -28,6 +28,12 @@ library;
 export 'src/digitraffic_advisory_provider.dart'
     show
         DigitrafficAdvisoryProvider,
+        DigitrafficCapMapping,
         DigitrafficHttpException,
         DigitrafficParseException,
+        defaultDigitrafficCapMapping,
+        defaultDigitrafficFallbackMapping,
+        kDefaultDigitrafficBoundingBoxHalfDegrees,
+        kDefaultDigitrafficTrafficAnnouncementsUrl,
+        kDigitrafficAttributionString,
         mapTrafficAnnouncementFeatureToAdvisory;